==> code/03.10_analysis_main_rf_ALE_forFig3_cropModel.Rmd <==
---
title: "formal analysis - random forest ALE, crop model data"
author: "jill deines"
date: "03/20/2022"
output: 
  html_document:
    toc: true
    toc_float: true
    keep_md: true
---

**Goal: use random forests to identify effects of WT using the modeled yield data**

Includes the sensitivity analysis/noisey simulations addition

```{r knitrOpts, echo=FALSE}
library(knitr)
opts_chunk$set(cache=FALSE, fig.path='../figure/03.03_analysis_main_rf_ale/')
```

**R Packages Needed**

```{r packages, message=FALSE, echo=TRUE, eval=TRUE}
library(dplyr)
library(ggplot2)
library(readr)
library(earth)
library(randomForest)

library(here)
library(patchwork)



sessionInfo()
```

# Directories

To run "analysis" scripts, please download the zipped data folder at [https://zenodo.org/doi/10.5281/zenodo.13274562](https://zenodo.org/doi/10.5281/zenodo.13274562) and save to your local computer. Use the filepath to "RepoData" for the `masterDataDir` variable at the top of each analysis script in the "Directories" code chunk.

```{r directories, message = FALSE}
# for master file
masterDataDir <- '/Users/Documents/projects/2022_compass_GLM/paper_gw_subsidy/data/RepoData'

# intermediate folder for large objects (eg rf output)
scratchFolder <- paste0(masterDataDir,'/intermediates')


# output folder: clean df for figures
repoDataDir <- paste0(here::here(),'/data/analysisOutput_forFigs')

```


# functions
helper functions for processing ALEplot output

```{r funs_ale}
library(ALEPlot)
aleToDf <- function(aleList, xname){
  df <- data.frame(xname = aleList$x.values,
                      'YieldEffect' = aleList$f.values)
  names(df)[1] <- xname
  return(df)
}

# prep prediction function for ALEplot
yhat <- function(X.model, newdata) as.numeric(predict(X.model, newdata, type = 'response'))

# for checking correlated variables
calcStats <- function(x_pred1, y_obs1){
  df <- data.frame(x_pred = x_pred1,
                   y_obs = y_obs1)
  # run a lm
  lm1 <- lm(y_obs ~ x_pred, data = df)

  # stats
  r2 <- summary(lm1)$r.squared
  rmse <- sqrt(mean((df$y_obs - df$x_pred)^2, na.rm = T))
  r <- cor(df$y_obs, df$x_pred)
  m <- coef(lm1)[2]
  int <- coef(lm1)[1]
  rmspe <- sqrt(mean(((df$y_obs - df$x_pred)/df$y_obs)^2))*100
  vecv <- (1 - sum((df$y_obs - df$x_pred)^2)/sum((df$y_obs - mean(df$y_obs))^2)) * 100
  mae <- sum(abs(df$y_obs-df$x_pred),na.rm = T)/nrow(df)
  mdae <- median(abs(df$y_obs-df$x_pred))
  me <- mean(df$x_pred-df$y_obs)

  # package into df
  outdf <- data.frame(r2 = r2, rmse = rmse, vecv = vecv, mae = mae, mdae = mdae, me = me,
                      r = r, m = m, int = int,
                      rmspe = rmspe)
  return(outdf)
}
```


# Crop Model (Enhanced with WT) model
with soil moisture from standard model

## random forest
run 1x and save output for later

```{r psim_rf, eval = FALSE}
# load enhanced model and standard model soil moisture
psimDiffs0 <- readRDS(paste0(masterDataDir,'/masterData_apsimGrid.rds'))
apsim_sm <- read_csv(paste0(masterDataDir,'/standardModel_JulySoilMoisture.csv'))

# filter years: same years as scym
psim <- psimDiffs0 %>% filter(year < 2019 & scym2020_10prct == 1) %>%
  dplyr::select(-contains('SW1')) %>% 
  left_join(apsim_sm)

# load MARS output for selected variables
marsAll <- readRDS( paste0(scratchFolder,'/marsOutput_scymDf_forVarSel.rds'))
ev <- earth::evimp(marsAll) # estimate variable importance
varsWanted <- rownames(ev)[1:17]
varsWanted2 <- varsWanted[!grepl('psim_07_m',varsWanted)]
varsWanted3 <- c(varsWanted2, 'WT_Jul_m','spei_ju1','SW1m_Jul')

# prep df for rf (enhanced model yield = psim_WT_tha)
psim_df  <- psim %>%
  dplyr::select(c(psim_WT_tha, all_of(varsWanted3) )) %>% tidyr::drop_na()
nrow(psim_df)

set.seed(5)
ncols <- dim(psim_df)[2]
rf_psim <- randomForest(x = psim_df[,2:ncols],
                         y = psim_df$psim_WT_tha,
                         ntree = 100,
                         importance = FALSE,
                        nodesize = 10)
rf_psim
saveRDS(rf_psim, paste0(scratchFolder,'/rf1_apsimGrid_marsVars_spei_WT_ns10_t100_May2024_standardSoilM.RDS'))
```

## psim ale plots

```{r psimAle}
# load enhanced model and standard model soil moisture
psimDiffs0 <- readRDS(paste0(masterDataDir,'/masterData_apsimGrid.rds'))
apsim_sm <- read_csv(paste0(masterDataDir,'/standardModel_JulySoilMoisture.csv'))

# filter years: same years as scym
psim <- psimDiffs0 %>% filter(year < 2019 & scym2020_10prct == 1) %>%
  dplyr::select(-contains('SW1')) %>% 
  left_join(apsim_sm)


# load MARS output for selected variables
marsAll <- readRDS( paste0(scratchFolder,'/marsOutput_scymDf_forVarSel.rds'))
ev <- earth::evimp(marsAll) # estimate variable importance
varsWanted <- rownames(ev)[1:17]
varsWanted2 <- varsWanted[!grepl('psim_07_m',varsWanted)]
varsWanted3 <- c(varsWanted2, 'WT_Jul_m','spei_ju1','SW1m_Jul')

# trim df for ale plot use
Xdat <- psim %>%
  dplyr::select(c(psim_WT_tha, psim_noWT_tha, all_of(varsWanted3) )) %>%
  tidyr::drop_na() %>%
  as.data.frame()

# model
rfModel <- readRDS( paste0(scratchFolder,'/rf1_apsimGrid_marsVars_spei_WT_ns10_t100_May2024_standardSoilM.RDS'))
rfModel

names(Xdat)

# psim - change with WT
ALEPlot_psim <- ALEPlot(X = Xdat, X.model = rfModel, J = 19, pred.fun = yhat, K = 100)
psimYield_df <- aleToDf(ALEPlot_psim, 'psim7')



write_csv(psimYield_df,paste0(repoDataDir,'/Fig3b_psim_yieldVsWT_aleplot_may2024_standardSoilM.csv'))

```


# Noisey simulation

based on these stack exchange posts 
* https://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variables.  
* and subsequent like https://stats.stackexchange.com/questions/38856/how-to-generate-correlated-random-numbers-given-means-variances-and-degree-of

Note: "deep" points at 4m mess up the correlation and are removed for this analysis

## generate noisy data

```{r noise}
psim %>%
  filter(WT_Jul_m < 3.9) %>%
  summarize(meanWT = mean(WT_Jul_m),
            stdev = sd(WT_Jul_m))

# add noise to wt data with correlation specified
# stack function 1
correlatedValue = function(x, r){
  r2 = r**2
  ve = 1-r2
  SD = sqrt(ve)
  e  = rnorm(length(x), mean=0, sd=SD)
  y  = r*x + e
  
  obsMean = mean(y)
  y2 = y - obsMean
  y2b = y2 * .6
  yOut = y2b + 1.9
  return(yOut)
}

# make correlated variable
psimShallow <- psim %>% filter(WT_Jul_m < 3.9) 
nrow(psim)
nrow(psimShallow)

psimShallow_df  <- psimShallow %>%
  dplyr::select(c(psim_WT_tha, all_of(varsWanted3) )) %>% tidyr::drop_na()

# function doesn't translate directly and has randomness - check values 
# that produce desired correlation; saved due to stochasticity

# psim_corVar_shallow <- psimShallow_df %>%
#   mutate(WT_Jul_cor40 = correlatedValue(WT_Jul_m, .57),
#          WT_Jul_cor50 = correlatedValue(WT_Jul_m,  .675),
#          WT_Jul_cor60 = correlatedValue(WT_Jul_m,  .76),
#          WT_Jul_cor70 = correlatedValue(WT_Jul_m,  .84),
#          WT_Jul_cor80 = correlatedValue(WT_Jul_m,  .9),
#          WT_psim = WT_Jul_m) %>%
#   rename(WT_Jul_cor100 = WT_Jul_m) %>%
#   tidyr::gather(., key = corrLevel, value = WT_Jul_m, contains('_cor')) 

psim_corVar_shallow <- read_csv(paste0(scratchFolder,'/df_corrVar_psim_WT_202405.csv'))

corAgreement <- psim_corVar_shallow %>%
  group_by(corrLevel) %>%
  do(as.data.frame(calcStats(.$WT_Jul_m, .$WT_psim))) %>%
  dplyr::select(corrLevel, r)
corAgreement
```

## run rf analysis on noisy data

```{r noise_rf, eval = FALSE}
psim_df1 <- psim_corVar_shallow %>% 
  dplyr::select(-c(WT_psim)) %>%
  filter(corrLevel == 'WT_Jul_cor60') %>%
  dplyr::select(-corrLevel)

set.seed(5)

ncols <- dim(psim_df1)[2]
rf_psim <- randomForest(x = psim_df1[,2:ncols],
                         y = psim_df1$psim_WT_tha,
                         ntree = 50,
                         importance = FALSE,
                        nodesize = 10)
rf_psim
saveRDS(rf_psim, paste0(scratchFolder,'/corrVar_psim_WT_ns10_t100_marsVars_noDeep_may24_WT_Jul_cor60.RDS'))

```

## noise ale plot

```{r noise_ale}
Xdat <- psim_corVar_shallow %>% 
  dplyr::select(-c(WT_psim)) %>%
  filter(corrLevel == 'WT_Jul_cor60') %>%
  dplyr::select(-c(psim_WT_tha, corrLevel)) %>%
  as.data.frame()

rfModel <- readRDS(paste0(scratchFolder,'/corrVar_psim_WT_ns10_t100_marsVars_noDeep_may24_WT_Jul_cor60.RDS'))

names(Xdat)

# psim - change with WT
ALEPlot_psim <- ALEPlot(X = Xdat, X.model = rfModel, J = 19, pred.fun = yhat, K = 100)

psimYield_df <- aleToDf(ALEPlot_psim, 'psim7') 

write_csv(psimYield_df, paste0(repoDataDir,'/Fig3b_psims_SA_noDeep_may24_cor60.csv'))  
```
